Removed controller & listeners for Interstitial
